refactoring with componentns
