+++ frontend/src/RenderDates.jsx
@@ -1,11 +1,23 @@
-export default function RenderDates(data) {
+export default function RenderDates({ data, prop, title }) {
+  const formatDate = function (stringDate) {
+    const options = {
+      year: "numeric",
+      month: "long",
+      day: "numeric",
+      weekday: "long",
+    };
+    const inputDate = new Date(stringDate);
+    return inputDate.toLocaleDateString("nl-NL", options);
+  };
+
+  const dates = data[prop];
+  const dateList =
+    dates && dates.map((date, idx) => <li key={idx}>{formatDate(date)}</li>);
 
   return (
-    <>
-      <h1>Yo this still works</h1>
-      <ul>
-        <Dates data={data} />
-      </ul>
-    </>
+    <div>
+      <h2>{title}</h2>
+      <ul>{dateList}</ul>
+    </div>
   );
 }
